Move reusable methods to a separate helper class
This way we allow other apps such as Geyser LocalSession to use these currently private methods without needing to copy over the code.
